Removes @tagless support
