+++ modules/internal/src/main/scala/server/fs2Calls.scala
@@ -20,8 +20,7 @@ package server
 
 import _root_.fs2.Stream
 import _root_.fs2.interop.reactivestreams._
-import cats.effect.{Effect, IO}
-import cats.syntax.functor._
+import cats.effect.Effect
 import io.grpc.stub.ServerCalls._
 import io.grpc.stub.StreamObserver
 import monix.execution.Scheduler
@@ -47,42 +46,26 @@ object fs2Calls {
       )
   }
 
-  def serverStreamingMethod[F[_], Req, Res](f: Req => F[Stream[F, Res]])(
+  def serverStreamingMethod[F[_], Req, Res](f: Req => Stream[F, Res])(
       implicit EFF: Effect[F],
       S: Scheduler): ServerStreamingMethod[Req, Res] = new ServerStreamingMethod[Req, Res] {
 
     override def invoke(request: Req, responseObserver: StreamObserver[Res]): Unit =
-      EFF
-        .runAsync(f(request)) {
-          case Right(stream) =>
-            println("1")
-
-            IO {
-              println("2")
-              Observable
-                .fromReactivePublisher(stream.toUnicastPublisher)
-                .subscribe(responseObserver)
-              println("3")
-            }.void
-          case Left(e) => IO.raiseError(e) // this will throw, but consistent previous impl
-        }
-        .unsafeRunSync
+      Observable
+        .fromReactivePublisher(f(request).toUnicastPublisher)
+        .subscribe(responseObserver)
   }
 
-  def bidiStreamingMethod[F[_], Req, Res](f: Stream[F, Req] => F[Stream[F, Res]])(
+  def bidiStreamingMethod[F[_], Req, Res](f: Stream[F, Req] => Stream[F, Res])(
       implicit EFF: Effect[F],
       S: Scheduler): BidiStreamingMethod[Req, Res] = new BidiStreamingMethod[Req, Res] {
 
     override def invoke(responseObserver: StreamObserver[Res]): StreamObserver[Req] =
       Subscriber2StreamObserver {
         monixCalls.transform[Req, Res](
-          (inputObservable: Observable[Req]) => {
-            Observable
-              .fromEffect(
-                f(inputObservable.toReactivePublisher.toStream[F])
-                  .map(ss => Observable.fromReactivePublisher(ss.toUnicastPublisher()))
-              )
-          }.flatten,
+          (inputObservable: Observable[Req]) =>
+            Observable.fromReactivePublisher(
+              f(inputObservable.toReactivePublisher.toStream[F]).toUnicastPublisher),
           StreamObserver2Subscriber(responseObserver)
         )
       }
