+++ modules/prometheus/server/src/test/scala/BaseMonitorServerInterceptorTests.scala
@@ -19,7 +19,7 @@ package prometheus
 package server
 
 import freestyle.rpc.common._
-import freestyle.rpc.withouttagless.Utils.client.MyRPCClient
+import freestyle.rpc.protocol.Utils.client.MyRPCClient
 import io.prometheus.client.Collector
 import freestyle.rpc.interceptors.metrics._
 
@@ -28,7 +28,7 @@ import scala.collection.JavaConverters._
 abstract class BaseMonitorServerInterceptorTests extends RpcBaseTestSuite {
 
   import freestyle.rpc.server.implicits._
-  import freestyle.rpc.withouttagless.Utils.database._
+  import freestyle.rpc.protocol.Utils.database._
   import freestyle.rpc.prometheus.shared.RegistryHelper._
 
   def name: String
