+++ modules/prometheus/server/src/test/scala/InterceptorsRuntime.scala
@@ -20,7 +20,7 @@ package server
 
 import freestyle.rpc.common.ConcurrentMonad
 import freestyle.rpc.prometheus.shared.Configuration
-import freestyle.rpc.withouttagless.Utils._
+import freestyle.rpc.protocol.Utils._
 import io.prometheus.client.CollectorRegistry
 
 case class InterceptorsRuntime(
